+++ src/neuroconv/datainterfaces/__init__.py
@@ -93,6 +93,7 @@ from .ophys.sima.simadatainterface import SimaSegmentationInterface
 from .ophys.suite2p.suite2pdatainterface import Suite2pSegmentationInterface
 from .ophys.tdt_fp.tdtfiberphotometrydatainterface import TDTFiberPhotometryInterface
 from .ophys.tiff.tiffdatainterface import TiffImagingInterface
+from .ophys.thor.thordatainterface import ThorImagingInterface
 
 # Image
 from .image.imageinterface import ImageInterface
@@ -158,6 +159,7 @@ interface_list = [
     MicroManagerTiffImagingInterface,
     MiniscopeImagingInterface,
     TDTFiberPhotometryInterface,
+    ThorImagingInterface,
     # Behavior
     VideoInterface,
     AudioInterface,